Create an Authorizator for single responsability
