+++ src/Api/Api.php
@@ -8,6 +8,7 @@ use kamermans\OAuth2\Persistence\FileTokenPersistence;
 use Namelivia\Fitbit\OAuth\Factory\Factory;
 use Namelivia\Fitbit\OAuth\Constants\Constants;
 use Namelivia\Fitbit\OAuth\Config\Config;
+use Namelivia\Fitbit\OAuth\Authorizator\Authorizator;
 
 class Api
 {
@@ -27,41 +28,8 @@ class Api
         $this->factory = $factory;
         $this->config = new Config($clientId, $clientSecret, $redirectUrl);
         $this->tokenPersistence = new FileTokenPersistence($tokenPath);
+        $this->authorizator = new Authorizator($this->config, $this->tokenPersistence);
         $this->initialized = false;
-    }
-
-    public function getAuthUri()
-    {
-        return Constants::AUTHORIZE_URL . '?' . http_build_query([
-            'client_id' => $this->config->getClientId(),
-            'scope' => implode(' ', [
-                'activity',
-                'nutrition',
-                'heartrate',
-                'location',
-                'nutrition',
-                'profile',
-                'settings',
-                'sleep',
-                'social',
-                'weight',
-            ]),
-            'response_type' => 'code',
-            'redirect_uri' => $this->config->getRedirectUrl(),
-            'expires_in' => '604800',
-        ]);
-    }
-
-    public function setAuthorizationCode(string $code)
-    {
-        $this->config->setCode($code);
-
-        return $this;
-    }
-
-    public function isAuthorized()
-    {
-        return $this->tokenPersistence->hasToken() || $this->config->hasCode();
     }
 
     public function isInitialized()
