+++ src/OAuth/Authorizator/Authorizator.php
@@ -1,0 +1,53 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Namelivia\Fitbit\OAuth\Authorizator;
+
+use kamermans\OAuth2\Persistence\FileTokenPersistence;
+use Namelivia\Fitbit\OAuth\Config\Config;
+
+class Authorizator
+{
+    public function __construct(
+        Config $config,
+        FileTokenPersistence $tokenPersistence
+    ) {
+        $this->config = $config;
+        $this->tokenPersistence = $tokenPersistence;
+    }
+
+    public function getAuthUri()
+    {
+        return Constants::AUTHORIZE_URL . '?' . http_build_query([
+            'client_id' => $this->config->getClientId(),
+            'scope' => implode(' ', [
+                'activity',
+                'nutrition',
+                'heartrate',
+                'location',
+                'nutrition',
+                'profile',
+                'settings',
+                'sleep',
+                'social',
+                'weight',
+            ]),
+            'response_type' => 'code',
+            'redirect_uri' => $this->config->getRedirectUrl(),
+            'expires_in' => '604800',
+        ]);
+    }
+
+    public function setAuthorizationCode(string $code)
+    {
+        $this->config->setCode($code);
+
+        return $this;
+    }
+
+    public function isAuthorized()
+    {
+        return $this->tokenPersistence->hasToken() || $this->config->hasCode();
+    }
+}
